RESTworld.AspNetCore 22.0.0 RESTworld.Business 14.1.0 RESTworld.Client.AspNetCore 15.0.0 RESTworld.Client.Net 7.1.0 RESTworld.Common 7.1.0 RESTworld.EntityFrameworkCore 7.1.0 RESTWorld.Testing 7.1.0 @wertzui/ngx-restworld-client 12.3.1
ICacheHelper now distinguishes between methods that will automatically insert the current user and ones that will not.
A bug causing tables in the frontend to not load correctly after the second time a table was rendered has been fixed.
Extra parameters for $filter, $orderby, $top, $skip are no longer required for ODataQueryOptions to work with OpenAPI and LinkGenerator.
Updated NuGet Packages